% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude.R
\docType{data}
\name{crude}
\alias{crude}
\title{20 Exemplary News Articles from the Reuters-21578 Data Set of Topic crude}
\format{
A VCorpus of 20 text documents.
source Reuters-21578 Text Categorization Collection Distribution 1.0 \acronym{XML} format
}
\usage{
data("crude")
}
\value{
Object class VCorpus
}
\description{
This data set holds 20 news articles with additional
meta information from the Reuters-21578 data set. All documents belong
to the topic \code{crude} dealing with crude oil
}
\examples{
\donttest{
data("crude")
crude}
}
\references{
Emms, Martin and Luz, Saturnino (2007). Machine Learning for Natural Language Processing.
\emph{European Summer School of Logic, Language and Information, course reader}.
}
\keyword{datasets}
